pyproject.toml
[tool.poetry]
name = "deposit-calculator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
psycopg2 = "^2.9.5"
pydantic = "^2.10.3"
python-dateutil = "^2.9.0.post0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
databases = "^0.9.0"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
